Split / refactor ScalaInterpreter
It was getting kind of hard to maintain
