+++ modules/scala/scala-interpreter/src/main/scala-2.13/almond/internals/ScalaInterpreterInspections.scala
@@ -1,5 +1,19 @@
 package almond.internals
 
-trait ScalaInterpreterInspections {
-  def inspectionsShutdown(): Unit = ()
+import almond.interpreter._
+import almond.logger.LoggerContext
+import ammonite.runtime.Frame
+import ammonite.util.Ref
+
+final class ScalaInterpreterInspections(
+  logCtx: LoggerContext,
+  metabrowse: Boolean,
+  metabrowseHost: String,
+  metabrowsePort: Int,
+  pressy: => scala.tools.nsc.interactive.Global,
+  frames: Ref[List[Frame]]
+) {
+  def inspect(code: String, pos: Int, detailLevel: Int): Option[Inspection] =
+    None
+  def shutdown(): Unit = ()
 }
